+++ internal/server/http/handlers/update/update_test.go
@@ -3,8 +3,8 @@ package update_test
 import (
 	"fmt"
 	"github.com/go-resty/resty/v2"
-	"github.com/mkolibaba/metrics/internal/http/router"
-	"github.com/mkolibaba/metrics/internal/storage"
+	"github.com/mkolibaba/metrics/internal/server/http/router"
+	"github.com/mkolibaba/metrics/internal/server/storage"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"net/http"
