+++ internal/server/http/router/router.go
@@ -2,10 +2,10 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/mkolibaba/metrics/internal/http/handlers/list"
-	"github.com/mkolibaba/metrics/internal/http/handlers/read"
-	"github.com/mkolibaba/metrics/internal/http/handlers/update"
-	"github.com/mkolibaba/metrics/internal/storage"
+	"github.com/mkolibaba/metrics/internal/server/http/handlers/list"
+	"github.com/mkolibaba/metrics/internal/server/http/handlers/read"
+	"github.com/mkolibaba/metrics/internal/server/http/handlers/update"
+	"github.com/mkolibaba/metrics/internal/server/storage"
 )
 
 func New(store storage.MetricsStorage) chi.Router {
